+++ src/ska_tmc_dishleafnode/commands/track_command.py
@@ -132,9 +132,12 @@ class Track(DishLNCommand):
             return result_code, message
 
         with self.component_manager.tango_operation_execution_lock:
+            self.logger.debug("Acquired  tango lock")
             result_code, message = self.call_adapter_method(
                 "Dish Master", self.dish_master_adapter, "Track"
             )
+
+        self.logger.debug("Released tango lock")
 
         if result_code[0] == ResultCode.FAILED:
             return result_code[0], message[0]