+++ tests/integration/test_dishln_configure_command.py
@@ -107,11 +107,14 @@ def configure_dish_leaf_node(
     tear_down(dish_leaf_node, dish_master, group_callback)
 
 
+# TO-Do - Enable non_sidereal_tracking once Error propogation is enabled.
+# We need to handle exception "Source is not visible currently" in test case.
+# @pytest.mark.parametrize(
+#     "json_to_use", ["dishleafnode_configure", "non_sidereal_tracking"]
+# )
 @pytest.mark.post_deployment
 @pytest.mark.SKA_mid
-@pytest.mark.parametrize(
-    "json_to_use", ["dishleafnode_configure", "non_sidereal_tracking"]
-)
+@pytest.mark.parametrize("json_to_use", ["dishleafnode_configure"])
 def test_configure_command(
     tango_context, group_callback, json_factory, json_to_use
 ):