+++ tests/unit/test_configure_command.py
@@ -57,7 +57,7 @@ def test_configure_command_completed(
 
     time.sleep(2)
     cm.update_device_pointing_state(PointingState.TRACK)
-    simulate_result_code_event(cm, "TRACK", ResultCode.OK)
+    simulate_result_code_event(cm, "Track", ResultCode.OK)
 
     task_callback.assert_against_call(
         call_kwargs={