+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+[0.18.0]
+********
+* Added DishLeafNode pointing tango device.
+
 [0.17.8]
 ********
 * Improved the timeout and error propagation for Configure command
@@ -12,8 +16,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 [0.17.3]
 ********
-* Fix bug SKB-502, to update the attributes at initialization, so that it can 
-show states of attributes at initialization on dashboards
+* Fix bug SKB-502, to update the attributes at initialization, so that it can show states of attributes at initialization on dashboards
 
 [0.17.2]
 ********
@@ -81,6 +84,15 @@ show states of attributes at initialization on dashboards
 
 Fixed
 ======
+
+[0.17.9]
+********
+* Error Propagation changes incorporated with process to stop track table.
+
+[0.17.7]
+********
+* Updated logic to stop program track table process.
+* Added Try Catch mechanism to identify issues faced in writing program track table on dish master
 
 
 [0.17.6]